+++ src/index.ts
@@ -1,15 +1,34 @@
-import { Middleware } from "rjweb-server"
+import { HttpRequestContext, Middleware } from "rjweb-server"
 import * as Sentry from "@sentry/node"
 
 /** @ts-ignore */
 import { version as pckgVersion } from "./pckg.json"
 export const version: string = pckgVersion
 
-export const sentry = new Middleware<Sentry.NodeOptions, { scope: Sentry.Scope }>('@rjweb/sentry', version)
+export class SentryHttpRequestContext extends HttpRequestContext {
+  /**
+   * Get the Sentry scope for the current request
+   * @since 2.1.0
+   * @from @rjweb/sentry
+  */ public scope(): Sentry.Scope {
+    return this.context.data(sentry).scope
+  }
+
+  /**
+   * Get the Sentry span for the current request
+   * @since 2.1.0
+   * @from @rjweb/sentry
+  */ public span(): Sentry.Span | null {
+    return this.context.data(sentry).span || null
+  }
+}
+
+export const sentry = new Middleware<Sentry.NodeOptions, { scope: Sentry.Scope, span?: Sentry.Span }>('@rjweb/sentry', version)
   .load((config) => {
     Sentry.init(config)
   })
-  .httpRequest((__, _, ctx, ctr) => {
+  .httpRequestContext(() => SentryHttpRequestContext)
+  .httpRequest(async(__, _, ctx, ctr) => {
     const scope = new Sentry.Scope()
     ctx.data(sentry).scope = scope
 
@@ -20,18 +39,18 @@ export const sentry = new Middleware<Sentry.NodeOptions, { scope: Sentry.Scope }
         queries: ctr.queries.toJSON(),
         fragments: ctr.fragments.toJSON(),
         route: ctx.route?.urlData.value?.toString() || null,
-        method: ctr.url.method,
-        body: ctr.rawBody || null
+        method: ctr.url.method
       })
-      .setSpan(Sentry.startTransaction({
-        op: ctr.type === 'http' ? 'http-request' : 'ws-upgrade',
-        name: `${ctr.type.toUpperCase()} ${ctr.url.method} ${ctx.route?.urlData.value?.toString() ?? '404'}`
-      }))
       .setUser({
         ip_address: ctr.client.ip.usual()
       })
 
-    const transaction = scope.getTransaction()!
+    const span = await new Promise<Sentry.Span | undefined>((resolve) => Sentry.startSpan({
+      name: `Request ${ctr.url.method} ${ctx.route?.urlData.value?.toString() ?? '404'}`,
+      op: ctr.type === 'http' ? 'http-request' : 'ws-upgrade'
+    }, (span) => resolve(span)))
+
+    if (span) ctx.data(sentry).span = span
 
     const handleError = ctx.handleError
     ctx.handleError = (err, cause) => {
@@ -43,15 +62,17 @@ export const sentry = new Middleware<Sentry.NodeOptions, { scope: Sentry.Scope }
       return original
     }
 
-    ctr.$abort(() => {
-      transaction.setHttpStatus(ctx.response.status)
-      transaction.finish()
+    if (span) ctr.$abort(() => {
+      Sentry.setHttpStatus(span, ctx.response.status)
+      span.end()
     })
   })
   .httpRequestFinish((_, __, ctx) => {
-    const transaction = ctx.data(sentry).scope.getTransaction()!
+    const span = ctx.data(sentry).span
 
-    transaction.setHttpStatus(ctx.response.status)
-    transaction.finish()
+    if (span) {
+      Sentry.setHttpStatus(span, ctx.response.status)
+      span.end()
+    }
   })
   .export()